requirements.txt
django==5.0.10
mysqlclient==2.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1

